soft of rewrote validate packet to allow for any mix of key value pairs. should also allow for command implementation

+++ boardcode/rpm.cpp
@@ -4,7 +4,7 @@
 RPMCollector::RPMCollector() {
 }
 
-auto RPMCollector::GetRPM(uint32_t timestamp, uint8_t pin_value) -> double {
+auto RPMCollector::GetRPM(int64_t timestamp, uint8_t pin_value) -> double {
     if (timestamp == 0) { return 0.0; }
     if (m_LastTime != 0 && timestamp > Thresh()) {
         m_RPM    = 0.0;
